Safari-611.1.10/LayoutTests/fast/forms/validation-message-minimum-font-size-expected.txt

Required text input:
Tests that the font used for validation bubble is aware of the minimumFontSize setting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS validationBubble.fontSize > 0 is true
PASS validationBubble.fontSize < 24 is true
PASS validationBubble.fontSize is 24
PASS successfullyParsed is true

TEST COMPLETE
